remove dependency on deck.gl for main example; use markers similar to interaction example instead
