+++ examples/main/data/bart-station.json
@@ -1,0 +1,46 @@
+[
+  {"name":"Lafayette (LAFY)","coordinates":[-122.123801,37.893394]},
+  {"name":"12th St. Oakland City Center (12TH)","coordinates":[-122.271604,37.803664]},
+  {"name":"16th St. Mission (16TH)","coordinates":[-122.419694,37.765062]},
+  {"name":"19th St. Oakland (19TH)","coordinates":[-122.269029,37.80787]},
+  {"name":"24th St. Mission (24TH)","coordinates":[-122.418466,37.752254]},
+  {"name":"Ashby (ASHB)","coordinates":[-122.26978,37.853024]},
+  {"name":"Balboa Park (BALB)","coordinates":[-122.447414,37.721981]},
+  {"name":"Bay Fair (BAYF)","coordinates":[-122.126871,37.697185]},
+  {"name":"Castro Valley (CAST)","coordinates":[-122.075567,37.690754]},
+  {"name":"Civic Center/UN Plaza (CIVC)","coordinates":[-122.413756,37.779528]},
+  {"name":"Colma (COLM)","coordinates":[-122.466233,37.684638]},
+  {"name":"Coliseum/Oakland Airport (COLS)","coordinates":[-122.197273,37.754006]},
+  {"name":"Concord (CONC)","coordinates":[-122.029095,37.973737]},
+  {"name":"Daly City (DALY)","coordinates":[-122.469081,37.706121]},
+  {"name":"Downtown Berkeley (DBRK)","coordinates":[-122.268045,37.869867]},
+  {"name":"El Cerrito del Norte (DELN)","coordinates":[-122.317269,37.925655]},
+  {"name":"Dublin/Pleasanton (DUBL)","coordinates":[-121.900367,37.701695]},
+  {"name":"Embarcadero (EMBR)","coordinates":[-122.396742,37.792976]},
+  {"name":"Fremont (FRMT)","coordinates":[-121.9764,37.557355]},
+  {"name":"Fruitvale (FTVL)","coordinates":[-122.224274,37.774963]},
+  {"name":"Glen Park (GLEN)","coordinates":[-122.434092,37.732921]},
+  {"name":"Hayward (HAYW)","coordinates":[-122.087967,37.670399]},
+  {"name":"Lake Merritt (LAKE)","coordinates":[-122.265609,37.797484]},
+  {"name":"MacArthur (MCAR)","coordinates":[-122.267227,37.828415]},
+  {"name":"Millbrae (MLBR)","coordinates":[-122.38666,37.599787]},
+  {"name":"Montgomery St. (MONT)","coordinates":[-122.401407,37.789256]},
+  {"name":"North Berkeley (NBRK)","coordinates":[-122.283451,37.87404]},
+  {"name":"North Concord/Martinez (NCON)","coordinates":[-122.024597,38.003275]},
+  {"name":"Orinda (ORIN)","coordinates":[-122.183791,37.878361]},
+  {"name":"Pleasant Hill/Contra Costa Centre (PHIL)","coordinates":[-122.056013,37.928403]},
+  {"name":"Pittsburg/Bay Point (PITT)","coordinates":[-121.945154,38.018914]},
+  {"name":"El Cerrito Plaza (PLZA)","coordinates":[-122.299272,37.903059]},
+  {"name":"Powell St. (POWL)","coordinates":[-122.406857,37.784991]},
+  {"name":"Richmond (RICH)","coordinates":[-122.353165,37.936887]},
+  {"name":"Rockridge (ROCK)","coordinates":[-122.251793,37.844601]},
+  {"name":"San Leandro (SANL)","coordinates":[-122.161311,37.722619]},
+  {"name":"San Bruno (SBRN)","coordinates":[-122.416038,37.637753]},
+  {"name":"San Francisco Int'l Airport (SFIA)","coordinates":[-122.392612,37.616035]},
+  {"name":"South Hayward (SHAY)","coordinates":[-122.057551,37.6348]},
+  {"name":"South San Francisco (SSAN)","coordinates":[-122.444116,37.664174]},
+  {"name":"Union City (UCTY)","coordinates":[-122.017867,37.591208]},
+  {"name":"Walnut Creek (WCRK)","coordinates":[-122.067423,37.905628]},
+  {"name":"West Dublin/Pleasanton (WDUB)","coordinates":[-121.928099,37.699759]},
+  {"name":"West Oakland (WOAK)","coordinates":[-122.294582,37.804675]}
+]
